extracted AbstractValidator
